Implement Turn Turtle (player_flip_camera)
This will create a new rendering camera, make it active, and render it.
Each tick it will follow the gameplay camera, except have it's roll set
to 180 degrees.

Currently, the biggest issue is that the camera will not go back to
normal once the effect ends. In my own mod, it works perfectly, but
inside of chaos something is different that I've yet to figure out. Any
input on this would be much appreciated.