Meta: Switch ChaosMod ASI mod to CMake.
Each time a new effect is committed, there's a lot of hoops to jump
through while merging these changes. The Visual Studio project/solution
is not fun to edit manually, isn't the greatest IDE, and makes building
more convoluted than needed sometimes.

CMake allows Visual Studio to still be used like normal! (the only thing
that will change is the C# projects being in the same solution, but that
doesn't matter). CMake can be used to generate a Visual Studio project
and solution to be used natively. More recent versions of Visual Studio
has better CMake support, but some may not be used to that.

For me, the improvement comes from non-Visual Studio build systems.
Ninja is a incredibly simple and fast build system that supports Windows
just fine. It also allows easier use of other IDEs like QtCreator or
CLion.

This commit creates three CMakeLists: one for the root, one for
ChaosMod, and one for MinHook.