Remove namespace prefix, revert Simeon Says changes.
The Simeon Says scaleform visibility time should use real-time. (it
should use real-time without pauses, but we don't have a metric for that
so this will have to do).

The real-time without pauses would be simple to implement, as it would
just take GetTickCount64 delta every time the main loop is ticked. While
paused, it doesn't tick.