ChaosMod: Add 'Effect Groups'
Fixes #667. Effects inside a specific group will have their weights adjusted in a way that should treat them as 'a single effect'. Currently this is applied to the majority of teleport, vehicle / object spawn and 'instant player kill' effects. A feature to allow users to customize those isn't planned (yet).