Added Effect "The Blade Hungers"
This is (sort of) from an escalation in Hitman 3 where 47 is given a katana and told "The Blade Hungers"

The screen flickers and slowly fades to red, and if it goes fully red you die. To reset the fade, you have to kill someone with the machete that the player is forced to hold.