ChaosMod/ConfigApp: Add option to disable effect group weighting
Since people seem to INSIST on not tampering with the weighting of specific effects the best course of action is to give them the option to disable that behavior completely. Hopefully people won't blindly disable it and then complain about the general balance of the mod, ugh.

Also fix weighting of meta effects which ignored the weight multiplier and used their own weight to increase their weight instead