Add "Google Translated" Effect
Replaces most of the labels in the game with text output which went through multiple translations. The corresponding hook will lazy cache the loaded labels to ensure they don't get loaded every time. This also pulls in another dependency for the used hash map since the one in the STL wasn't up to speed.