Made improvements.
You can now grab AI vehicles. You can now hit the shoot button to send an object flying through the air. Changed to RayCasting from the gameplay cam to get entities as they wouldn't get picked up properly sometimes with the old method. Switched to detecting input instead of getting the players aiming state. You no longer have to be constantly aiming to pickup an object. You press it, it scans, if there is an object it will pick it up for as long as you have he aim button down. Thanks to pongo for helping me with some errors. And MoneyWasted for the RayCasting.