ChaosMod: Fix player switch effects with ScriptThreadBlock (#3350)
Previously you could hold the character switch button, then activate an effect that uses ScriptThreadBlock (an example would be Blimp Strats). And the wheel would be disabled, but the screen effects and slow-mo would still be active. That would be the case until DisableScriptThreadBlock is called.

This fixes that by setting the time scale to full and stopping all screen effects that frame.