ChaosMod: Implement debug shortcuts (#2996)
#### If the build configuration is set to debug:
You can set a debug shortcut for effects with `.DebugShortcutKeycode` in
EffectInfo. This will override shortcuts read from the config file (if
any), with the specified KeyCode. This saves time having to scroll
through the debug menu. I was originally going to do the same with the
lua scripts (making the member only functional when using a debug
build), but people writing lua scripts, don't necessarily have a debug
build.

NOTE: The member doesn't exist in Release configuration.

---------

Co-authored-by: pongo1231 <pongo1999712@gmail.com>