Add "ESP" Effect (#3432)
This effect implements an ESP (Extra Sensory Perception), which is a
common cheat in games that allows the player to see other players or
NPCs behind solid geometry. This specific ESP draws the skeleton of
pedestrians within a certain distance from the player and with a random
line color.

There are some technical limitations, such as the ESP being drawn one
frame behind the rest of the game. This is due to how Rockstar handles
rendering. This can be solved by using other methods of rendering, such
as hooking DirectX directly to draw the lines with a shader. This was my
first approach as it would not cause this issue, but it became too much
of a hassle.

The current method of how the line is rendered is by hooking the same
function that is responsible for the other drawing natives, such as
DRAW_RECT. Doing this allows me to reuse existing code that Rockstar has
already developed.

Below is what the ESP looks like on a pedestrian.
![Skeleton Drawing
Example](https://user-images.githubusercontent.com/45946092/233813673-7c662852-9674-43e0-8ced-8f28a3c88127.PNG)

---------

Co-authored-by: pongo1231 <pongo1999712@gmail.com>